fix: add new files structure. Move actions implementations into separate modules.
