+++ nwd/ls.js
@@ -1,7 +1,6 @@
-import { commandsEmitter } from "./emitter.js";
-import { access, readdir, constants } from "node:fs/promises";
-import store from "./store.js";
-import path from "node:path";
+import { readdir } from "node:fs/promises";
+
+import store from "./../store.js";
 
 let currentDir = "";
 
@@ -9,32 +8,11 @@ store.onUpdate((state) => {
   currentDir = state.workingDirectory;
 });
 
-commandsEmitter.on("up", () => {
-  store.trigger({ type: "SET_WORKING_DIR", payload: path.dirname(currentDir) });
-});
-
-commandsEmitter.on("cd", async (args) => {
-  const [changeDirectoryTo] = args;
-
-  let newDir = changeDirectoryTo;
-  if (!path.isAbsolute(changeDirectoryTo)) {
-    newDir = path.resolve(currentDir, changeDirectoryTo);
-  }
-
-  try {
-    await access(newDir, constants.F_OK);
-  } catch (error) {
-    console.error("Invalid input");
-    return;
-  }
-
-  store.trigger({ type: "SET_WORKING_DIR", payload: newDir });
-});
-
 const sortEntriesAlphabetically = (arr) => {
   return [...arr].sort((a, b) => a.Name.localeCompare(b.Name));
 };
-commandsEmitter.on("ls", async () => {
+
+export const ls = async () => {
   try {
     const directoryEntries = await readdir(currentDir, { withFileTypes: true });
     const { directories, files } = directoryEntries.reduce(
@@ -75,4 +53,4 @@ commandsEmitter.on("ls", async () => {
       throw new Error("FS operation failed");
     }
   }
-});
+};
